Fix physical Escape RmlUi input path
Record SDL keyboard transitions before modal consumption, latch Escape edges, and extract keyboard state into the input subsystem. Add a real XTEST route, Clang build/format/tidy gates, and Python project tooling.

+++ runtime/ui/rmlui_backend.cpp
@@ -3,11 +3,9 @@
 
 RmlUiRenderBackend::RmlUiRenderBackend(SDL_Renderer *renderer) : renderer(renderer)
 {
-    blend = SDL_ComposeCustomBlendMode(SDL_BLENDFACTOR_ONE,
-                                       SDL_BLENDFACTOR_ONE_MINUS_SRC_ALPHA,
-                                       SDL_BLENDOPERATION_ADD, SDL_BLENDFACTOR_ONE,
-                                       SDL_BLENDFACTOR_ONE_MINUS_SRC_ALPHA,
-                                       SDL_BLENDOPERATION_ADD);
+    blend =
+        SDL_ComposeCustomBlendMode(SDL_BLENDFACTOR_ONE, SDL_BLENDFACTOR_ONE_MINUS_SRC_ALPHA, SDL_BLENDOPERATION_ADD,
+                                   SDL_BLENDFACTOR_ONE, SDL_BLENDFACTOR_ONE_MINUS_SRC_ALPHA, SDL_BLENDOPERATION_ADD);
 }
 
 void RmlUiRenderBackend::BeginFrame()
@@ -29,33 +27,30 @@ void RmlUiRenderBackend::EndFrame()
     scissor_enabled = false;
 }
 
-Rml::CompiledGeometryHandle RmlUiRenderBackend::CompileGeometry(
-    Rml::Span<const Rml::Vertex> source_vertices, Rml::Span<const int> source_indices)
+Rml::CompiledGeometryHandle RmlUiRenderBackend::CompileGeometry(Rml::Span<const Rml::Vertex> source_vertices,
+                                                                Rml::Span<const int> source_indices)
 {
     auto geometry = std::make_unique<Geometry>();
     geometry->vertex_count = (int)source_vertices.size();
     geometry->vertices.reset(new SDL_Vertex[geometry->vertex_count]);
     for (int i = 0; i < geometry->vertex_count; i++) {
         const Rml::Vertex &v = source_vertices[i];
-        geometry->vertices[i].position = { v.position.x, v.position.y };
-        geometry->vertices[i].tex_coord = { v.tex_coord.x, v.tex_coord.y };
-        geometry->vertices[i].color = { v.colour.red / 255.f, v.colour.green / 255.f,
-                                        v.colour.blue / 255.f, v.colour.alpha / 255.f };
+        geometry->vertices[i].position = {v.position.x, v.position.y};
+        geometry->vertices[i].tex_coord = {v.tex_coord.x, v.tex_coord.y};
+        geometry->vertices[i].color = {
+            static_cast<float>(v.colour.red) / 255.f, static_cast<float>(v.colour.green) / 255.f,
+            static_cast<float>(v.colour.blue) / 255.f, static_cast<float>(v.colour.alpha) / 255.f};
     }
     geometry->index_count = (int)source_indices.size();
     geometry->indices.reset(new int[geometry->index_count]);
-    std::memcpy(geometry->indices.get(), source_indices.data(),
-                (size_t)geometry->index_count * sizeof(int));
+    std::memcpy(geometry->indices.get(), source_indices.data(), (size_t)geometry->index_count * sizeof(int));
     return reinterpret_cast<Rml::CompiledGeometryHandle>(geometry.release());
 }
 
 void RmlUiRenderBackend::ReleaseGeometry(Rml::CompiledGeometryHandle handle)
-{
-    delete reinterpret_cast<Geometry *>(handle);
-}
+{ delete reinterpret_cast<Geometry *>(handle); }
 
-void RmlUiRenderBackend::RenderGeometry(Rml::CompiledGeometryHandle handle,
-                                        Rml::Vector2f translation,
+void RmlUiRenderBackend::RenderGeometry(Rml::CompiledGeometryHandle handle, Rml::Vector2f translation,
                                         Rml::TextureHandle texture)
 {
     Geometry *geometry = reinterpret_cast<Geometry *>(handle);
@@ -65,12 +60,11 @@ void RmlUiRenderBackend::RenderGeometry(Rml::CompiledGeometryHandle handle,
         vertices[i].position.x += translation.x;
         vertices[i].position.y += translation.y;
     }
-    SDL_RenderGeometry(renderer, reinterpret_cast<SDL_Texture *>(texture), vertices.get(),
-                       geometry->vertex_count, geometry->indices.get(), geometry->index_count);
+    SDL_RenderGeometry(renderer, reinterpret_cast<SDL_Texture *>(texture), vertices.get(), geometry->vertex_count,
+                       geometry->indices.get(), geometry->index_count);
 }
 
-Rml::TextureHandle RmlUiRenderBackend::LoadTexture(Rml::Vector2i &dimensions,
-                                                   const Rml::String &source)
+Rml::TextureHandle RmlUiRenderBackend::LoadTexture(Rml::Vector2i &dimensions, const Rml::String &source)
 {
     const DeviceAsset asset = device_asset_from_source(source.c_str());
     if (asset == DEVICE_ASSET_INVALID) return 0;
@@ -88,7 +82,7 @@ Rml::TextureHandle RmlUiRenderBackend::LoadTexture(Rml::Vector2i &dimensions,
         return 0;
     }
 
-    dimensions = { surface->w, surface->h };
+    dimensions = {surface->w, surface->h};
     SDL_Texture *texture = SDL_CreateTextureFromSurface(renderer, surface);
     SDL_DestroySurface(surface);
     if (texture) {
@@ -102,12 +96,10 @@ Rml::TextureHandle RmlUiRenderBackend::LoadTexture(Rml::Vector2i &dimensions,
     return reinterpret_cast<Rml::TextureHandle>(texture);
 }
 
-Rml::TextureHandle RmlUiRenderBackend::GenerateTexture(Rml::Span<const Rml::byte> source,
-                                                       Rml::Vector2i dimensions)
+Rml::TextureHandle RmlUiRenderBackend::GenerateTexture(Rml::Span<const Rml::byte> source, Rml::Vector2i dimensions)
 {
-    SDL_Texture *texture = SDL_CreateTexture(renderer, SDL_PIXELFORMAT_RGBA32,
-                                             SDL_TEXTUREACCESS_STATIC,
-                                             dimensions.x, dimensions.y);
+    SDL_Texture *texture =
+        SDL_CreateTexture(renderer, SDL_PIXELFORMAT_RGBA32, SDL_TEXTUREACCESS_STATIC, dimensions.x, dimensions.y);
     if (!texture) return 0;
     SDL_SetTextureBlendMode(texture, blend);
     /* SDL3 returns true on success. Treating it like SDL2's integer status destroyed every
@@ -132,7 +124,9 @@ void RmlUiRenderBackend::EnableScissorRegion(bool enable)
 
 void RmlUiRenderBackend::SetScissorRegion(Rml::Rectanglei region)
 {
-    scissor.x = region.Left(); scissor.y = region.Top();
-    scissor.w = region.Width(); scissor.h = region.Height();
+    scissor.x = region.Left();
+    scissor.y = region.Top();
+    scissor.w = region.Width();
+    scissor.h = region.Height();
     if (scissor_enabled) SDL_SetRenderClipRect(renderer, &scissor);
 }
